Remove `blah` which is not a descriptive variable name.
cc: @felixcodes.  Please make sure to clean up stuff like this before asking me or other people to review :) 

Normally I won't fix this for you.  However, in this case I also want to show you how to properly format chained methods.  
Basically always put function's content on its own line.  Otherwise, we can't add a breakpoint in debugging.   
(If you find that we violate this elsewhere in the codebase, feel free to refactor.)